refactor: move font manager alongside Ayah components
