+++ src/lib/components/Quran/FontManager.svelte.ts
